Final Project AKA Project 4
Not to spoil anything but here is cheat code to winning this game (Just
so you don't get confused to win the game you must follow the map and go
to location 2 and take the knife and then go to gym and drink the
protein shake to win the game. So you go east then east again to the
storage room then press "t"  to take the knife and then head over to
location 5 which is the gym which is south from the storage room then
you press yes to drink it, then you  win the game!!! I tried my best
make it perfect I really hope it was enough! )